+++ __tests__/getIso4217Currencies.test.ts
@@ -1,10 +1,8 @@
-import { getIso4217Currencies, Iso4217Currency } from "../src";
+import { getIso4217Currencies } from "../src";
 import { iso4217Currencies } from "../src/internal/iso4217Currencies";
 
 const isArrayUnique = (arr: any[]) => Array.isArray(arr) 
   && new Set(arr).size === arr.length;
-
-const nameof = <T>(name: keyof T) => name;
 
 describe('getIso4217Currencies', () => {
   test('all properties are properly defined', () => {
@@ -24,10 +22,23 @@ describe('getIso4217Currencies', () => {
     expect(countries.find(x => x.alpha3Code === "HRK")).toBeUndefined();
   });
 
-  test('Kuna currency is returned', () => {
+  test('Kuna (HRK) currency is returned', () => {
     const countries = getIso4217Currencies("2022-12-31");
     expect(countries).toHaveLength(181);
     expect(countries.find(x => x.alpha3Code === "HRK")).toBeDefined();
+  });
+
+  test('Caribbean Guilder (XCG) and Netherlands Antillean Guilder (ANG) currencies are returned', () => {
+    const countries = getIso4217Currencies("2025-03-31");
+    expect(countries).toHaveLength(181);
+    expect(countries.find(x => x.alpha3Code === "XCG")).toBeDefined();
+    expect(countries.find(x => x.alpha3Code === "ANG")).toBeDefined();
+  });
+
+  test('Netherlands Antillean Guilder (ANG) currency is no longer returned', () => {
+    const countries = getIso4217Currencies("2025-07-01");
+    expect(countries).toHaveLength(180);
+    expect(countries.find(x => x.alpha3Code === "ANG")).toBeUndefined();
   });
 
   test('function should return the same reference', () => {
@@ -36,10 +47,14 @@ describe('getIso4217Currencies', () => {
     expect(countries1 === countries2).toBeTruthy();
   });
 
-  test.each([
-    [nameof<Iso4217Currency>("alpha3Code")],
-    [nameof<Iso4217Currency>("numericCode")],
-  ])('no duplicates for property %p', (propName) => {
-    expect(isArrayUnique(iso4217Currencies.map(x => x[propName]))).toBeTruthy();
+  test('no duplicates for property alpha3Code', () => {
+    expect(isArrayUnique(iso4217Currencies.map(x => x.alpha3Code))).toBeTruthy();
+  });
+
+  test('no duplicates for property numericCode', () => {
+    const currenciesWithoutKnownNumericCodeDuplicates = iso4217Currencies
+      .filter(x => x.alpha3Code !== "ANG" && x.alpha3Code !== "XCG");
+
+    expect(isArrayUnique(currenciesWithoutKnownNumericCodeDuplicates.map(x => x.numericCode))).toBeTruthy();
   });
 });