+++ src/findIso4217CurrencyForIso3166Country.ts
@@ -1,10 +1,14 @@
+import { convertToDate } from "./internal/convertToDate";
 import { iso3166Countries } from "./internal/iso3166Countries";
 import { iso3166CountryToIso4217Currency } from "./internal/iso3166CountryToIso4217Currency";
 import { iso4217Currencies } from "./internal/iso4217Currencies";
+import { DateIsoString } from "./types/dateIsoString";
 import { Iso3166Alpha2Code, Iso3166Alpha3Code, Iso3166NumericCode } from "./types/iso3166";
 import { Iso4217Currency } from "./types/iso4217";
 
-export function findIso4217CurrencyForIso3166Country(code: Iso3166Alpha2Code | Iso3166Alpha3Code | Iso3166NumericCode): Iso4217Currency | undefined {
+export function findIso4217CurrencyForIso3166Country(
+  code: Iso3166Alpha2Code | Iso3166Alpha3Code | Iso3166NumericCode,
+  statusForTheDay?: Date | DateIsoString): Iso4217Currency | undefined {
   const countryAlpha3Code = typeof code === "string"
     ? code.length === 3
       ? code
@@ -15,11 +19,18 @@ export function findIso4217CurrencyForIso3166Country(code: Iso3166Alpha2Code | I
     return undefined;
   }
 
-  const currencyAlpha3Code = iso3166CountryToIso4217Currency.get(countryAlpha3Code);
+  const resolver = iso3166CountryToIso4217Currency.get(countryAlpha3Code);
 
-  if (!currencyAlpha3Code) {
+  if (!resolver) {
     return undefined;
   }
 
-  return iso4217Currencies.find(x => x.alpha3Code === currencyAlpha3Code);
+  if (typeof resolver === "string") {
+    return iso4217Currencies.find(x => x.alpha3Code === resolver);
+  }
+
+  const pointInTime = statusForTheDay ? convertToDate(statusForTheDay) : new Date();
+  const resolvedCurrencyAlpha3CodeByDate = resolver(pointInTime);
+
+  return iso4217Currencies.find(x => x.alpha3Code === resolvedCurrencyAlpha3CodeByDate);
 }